requirements.txt
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
billiard==3.3.0.20
celery==3.1.17
Flask==0.10.1
Flask-RESTful==0.3.3
Flask-SQLAlchemy==2.0
iso8601==0.1.10
itsdangerous==0.24
Jinja2==2.7.3
kombu==3.0.26
MarkupSafe==0.23
msgpack-python==0.4.6
netaddr==0.7.14
netifaces==0.10.4
oslo.config==1.11.0
oslo.i18n==1.6.0
oslo.serialization==1.5.0
oslo.utils==1.5.0
pbr==1.0.1
prettytable==0.7.2
python-keystoneclient==1.5.0
python-novaclient==2.25.0
pytz==2015.4
redis==2.10.3
requests==2.7.0
simplejson==3.7.3
six==1.9.0
SQLAlchemy==1.0.5
stevedore==1.4.0
Werkzeug==0.10.4
wheel==0.24.0
PyYAML==3.11
Flask-Mail==0.9.1
Flask-Script==2.0.5
python3-ldap==0.9.8.4
psycopg2==2.6.1
Flask-Testing==0.4.2
